ai-engine/requirements.txt
tensorflow==2.20.0
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
requests==2.31.0
python-dotenv==1.0.0
web3==6.9.0
fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.3.0
aiohttp==3.8.5
asyncio-mqtt==0.13.0
schedule==1.2.0
ta==0.10.2
yfinance==0.2.18
ccxt==4.0.77
tweepy==4.14.0
textblob==0.17.1
redis==4.6.0
psycopg2-binary==2.9.7
sqlalchemy==2.0.20
alembic==1.12.0
cryptography==41.0.4
eth-account==0.9.0
websockets==11.0.3
python-multipart==0.0.6
httpx==0.24.1
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.15.0
jupyter==1.0.0
ipykernel==6.25.1
